requirements.txt
alembic==1.14.0
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
celery-stubs==0.1.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.48
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.36
sqlalchemy-stubs==0.4
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
